benchmark/requirements.txt
snowflake-connector-python>=3.0.0
python-dotenv>=1.0.0
tabulate>=0.9.0
pandas>=1.5.0
matplotlib>=3.5.0
numpy>=1.21.0
requests>=2.25.0